Moved the error check to right after mvn

+++ scripts/linux/oldmar-build.sh
@@ -14,6 +14,10 @@ if [ ! -f $OSSIM_INSTALL_PREFIX/share/java/joms-$OSSIM_VERSION.jar ]; then
 fi
 # make sure the joms jar is in the local maven repo
 mvn install:install-file -Dfile=$OSSIM_DEV_HOME/install/share/java/joms-$OSSIM_VERSION.jar -DgroupId=org.ossim -DartifactId=joms -Dversion=$OSSIM_VERSION -Dpackaging=jar
+if [ $? -ne 0 ]; then
+ echo; echo "ERROR: MVN install failed for joms."
+ exit 1
+fi
 
 if [ -d $OSSIM_DEV_HOME/ossim-install ] ; then
    rm -rf $OSSIM_DEV_HOME/ossim-install
@@ -21,10 +25,6 @@ fi
 
 mv $OSSIM_INSTALL_PREFIX ossim-install
 
-if [ $? -ne 0 ]; then
- echo; echo "ERROR: MVN install failed for joms."
- exit 1
-fi
 
 $OSSIM_DEV_HOME/omar/build_scripts/linux/build.sh
 if [ $? -ne 0 ]; then
